requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
pyyaml==6.0.2
pyarrow==17.0.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
torch==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.4
streamlit==1.37.0
